requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
auth0-python==4.4.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.4
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-redis==5.3.0
django-timezone-field==6.0.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
idna==3.4
kombu==5.3.2
multidict==6.0.4
mysqlclient==2.2.0
phonenumbers==8.13.21
Pillow==10.0.1
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
